pyproject.toml
[project]
name = "app-store-connect-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.4",
    "httpx>=0.27",
    "python-dotenv>=1.0.1",
    "email-validator>=2.1.0",
    "pyjwt>=2.10.1",
    "cryptography>=45.0.7",
]

[project.scripts]
app-store-connect-mcp = "app_store_connect_mcp.server:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]


[project.optional-dependencies]
dev = [
    "datamodel-code-generator>=0.25.7",
    "pytest>=8.3.0",
    "pytest-asyncio>=1.1.0",
]
